+++ generators/java/generator-utils/src/main/java/com/fern/java/generators/AliasGenerator.java
@@ -286,7 +286,7 @@ public final class AliasGenerator extends AbstractTypeGenerator {
 
         @Override
         public CodeBlock visitUint64() {
-            return CodeBlock.of("return $T.$L($L)", Integer.class, "toString", VALUE_FIELD_NAME);
+            return CodeBlock.of("return $T.$L($L)", Long.class, "toString", VALUE_FIELD_NAME);
         }
 
         @Override
